+++ core/src/test/java/io/github/lemon_ant/jharmonizer/core/files_handler/SrcFileCreator.java
@@ -8,6 +8,7 @@ import lombok.NonNull;
 public class SrcFileCreator {
 
     @NonNull
+    // TODO Change parameters order
     public static SrcFile createSrcFile(@NonNull String srcCode, @NonNull Path path) {
         return new SrcFile(path, srcCode);
     }